ArrayWrapper.save(npz)
