+++ openquake/server/utils.py
@@ -133,38 +133,3 @@ def check_webserver_running(url="http://localhost:8800", max_retries=30):
         logging.warning('Unable to connect to %s within %s retries'
                         % (url, max_retries))
     return success
-
-
-def fix_array(arr, key):
-    """
-    :param arr: array or array-like object
-    :param key: string associated to the error (appear in the error message)
-
-    If `arr` is a numpy array with dtype object containing strings, convert
-    it into a numpy array containing bytes, unless it has more than 2
-    dimensions or contains non-strings (these are errors). Return `arr`
-    unchanged in the other cases.
-    """
-    if arr is None:
-        return ()
-    if not isinstance(arr, numpy.ndarray):
-        return arr
-    if arr.dtype != numpy.dtype('O'):
-        d = arr.dtype.descr
-        if len(d) > 1 and isinstance(d[0][1], tuple):
-            # for extract_assets d[0] is the pair
-            # ('id', ('|S20', {'h5py_encoding': 'ascii'}))
-            # this is a horrible workaround for the h5py 2.10.0 issue
-            # https://github.com/numpy/numpy/issues/14142#issuecomment-620980980
-            arr.dtype = [(n, str(arr.dtype[n])) for n in arr.dtype.names]
-        return arr
-    if arr.ndim == 1:
-        return numpy.array([s.encode('utf8') for s in arr])
-    elif arr.ndim == 2:
-        return numpy.array([[col.encode('utf8') for col in row]
-                            for row in arr])
-    else:
-        raise NotImplementedError('The array for %s has shape %s' %
-                                  (key, arr.shape))
-
-    return arr
